rtc: omap: Add runtime pm support
OMAP1 RTC driver is used in multiple devices like,
OMAPL138 and AM33XX. Driver currently doesn't handle any clocks,
which may be right for OMAP1 architecture but in case of AM33XX,
the clock/module needs to be enabled in order to access the registers.

So covert this driver to runtime pm, which internally handles rest.

Signed-off-by: Vaibhav Hiremath <hvaibhav@ti.com>

This patch has been tested on OMAPL138, BeagleBone platform,
help is required to validate it on OMAP1 platform.

Also note that, suspend/resume execution patch has not been validated,
but theoretically should work, since the driver is tested with building
as a module; where it has been made sure that, on insmode clock/module
is enabled and on rmmod clock/module is disabled, without impacting
rtc functionality.