rcu: Move unanticipated condition from "if" to "WARN_ON()"
The rcu_preempt_offline_tasks() checks for any of the rcu_node structure's
CPUs still being in an RCU critical section. However, it is extremely
unlikely that any will be due to the fact that all the CPUs must be
offline when this function is called.  This commit therefore moves the
check from the "if" statement to a "WARN_ON()".

Signed-off-by: Paul E. McKenney <paul.mckenney@linaro.org>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>