ARM: OMAP2+: Prepare for irqs.h removal
As the interrupts should only be defined in the platform_data, and
eventually coming from device tree, there's no need to define them
in header files.

Let's remove the hardcoded references to irqs.h and fix up the includes
so we don't rely on headers included in irqs.h. Note that we're
defining OMAP_INTC_START as 0 to the interrupts. This will be needed
when we enable SPARSE_IRQ.

While at it, sort the includes the standard way, and add the trailing
commas where they are missing in the related data structures.

Note that for drivers/staging/tidspbridge we just define things
locally.

Signed-off-by: Tony Lindgren <tony@atomide.com>