mm/compaction: Move migration fail/success stats to migrate.c
The compact_pages_moved and compact_pagemigrate_failed events
are convenient for determining if compaction is active and to
what degree migration is succeeding but it's at the wrong level.
Other users of migration may also want to know if migration is
working properly and this will be particularly true for any
automated NUMA migration. This patch moves the counters down to
migration with the new events called pgmigrate_success and
pgmigrate_fail. The compact_blocks_moved counter is removed
because while it was useful for debugging initially, it's
worthless now as no meaningful conclusions can be drawn from its
value.

Signed-off-by: Mel Gorman <mgorman@suse.de>
Reviewed-by: Rik van Riel <riel@redhat.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Hugh Dickins <hughd@google.com>
Cc: Paul Turner <pjt@google.com>
Cc: Lee Schermerhorn <Lee.Schermerhorn@hp.com>
Cc: Alex Shi <lkml.alex@gmail.com>
Cc: Srikar Dronamraju <srikar@linux.vnet.ibm.com>
Cc: Aneesh Kumar <aneesh.kumar@linux.vnet.ibm.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Peter Zijlstra <a.p.zijlstra@chello.nl>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Ingo Molnar <mingo@kernel.org>