mm/mempolicy: Add MPOL_MF_NOOP
NOTE: I have not yet addressed by own review feedback of this
patch. At 	this point I'm trying to construct a baseline tree and will apply
	my own review feedback later and then fold it in.

This patch augments the MPOL_MF_LAZY feature by adding a "NOOP"
policy to mbind().  When the NOOP policy is used with the 'MOVE
and 'LAZY flags, mbind() will map the pages PROT_NONE so that
they will be migrated on the next touch.

This allows an application to prepare for a new phase of
operation where different regions of shared storage will be
assigned to worker threads, w/o changing policy.  Note that we
could just use "default" policy in this case.  However, this
also allows an application to request that pages be migrated,
only if necessary, to follow any arbitrary policy that might
currently apply to a range of pages, without knowing the policy,
or without specifying multiple mbind()s for ranges with
different policies.

[ Bug in early version of mpol_parse_str() reported by Fengguang Wu. ]

Reported-by: Fengguang Wu <fengguang.wu@intel.com>
Signed-off-by: Lee Schermerhorn <lee.schermerhorn@hp.com>
Reviewed-by: Rik van Riel <riel@redhat.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Hugh Dickins <hughd@google.com>
Cc: Paul Turner <pjt@google.com>
Cc: Lee Schermerhorn <Lee.Schermerhorn@hp.com>
Cc: Alex Shi <lkml.alex@gmail.com>
Cc: Srikar Dronamraju <srikar@linux.vnet.ibm.com>
Cc: Aneesh Kumar <aneesh.kumar@linux.vnet.ibm.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Andrew Morton <akpm@linux-foundation.org>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Peter Zijlstra <a.p.zijlstra@chello.nl>
Signed-off-by: Mel Gorman <mgorman@suse.de>
Signed-off-by: Ingo Molnar <mingo@kernel.org>