mm/mempolicy: Add MPOL_MF_LAZY
NOTE: Once again there is a lot of patch reuse and the end
      result is sufficiently different that I had to drop
      the prior signed-offs.
      Will re-add if the original authors are ok with that.

This patch adds another mbind() flag to request "lazy
migration".  The flag, MPOL_MF_LAZY, modifies MPOL_MF_MOVE* such
that the selected pages are marked PROT_NONE. The pages will be
migrated in the fault path on "first touch", if the policy
dictates at that time.

"Lazy Migration" will allow testing of migrate-on-fault via
mbind(). Also allows applications to specify that only
subsequently touched pages be migrated to obey new policy,
instead of all pages in range. This can be useful for
multi-threaded applications working on a large shared data area
that is initialized by an initial thread resulting in all pages
on one [or a few, if overflowed] nodes. After PROT_NONE, the
pages in regions assigned to the worker threads will be
automatically migrated local to the threads on 1st touch.

Signed-off-by: Mel Gorman <mgorman@suse.de>
Reviewed-by: Rik van Riel <riel@redhat.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Hugh Dickins <hughd@google.com>
Cc: Paul Turner <pjt@google.com>
Cc: Lee Schermerhorn <Lee.Schermerhorn@hp.com>
Cc: Alex Shi <lkml.alex@gmail.com>
Cc: Srikar Dronamraju <srikar@linux.vnet.ibm.com>
Cc: Aneesh Kumar <aneesh.kumar@linux.vnet.ibm.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Peter Zijlstra <a.p.zijlstra@chello.nl>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Ingo Molnar <mingo@kernel.org>