dtc: Dynamic symbols & fixup support
Enable the generation of symbol & fixup information for
usage with dynamic DT loading.

Passing the -@ option generates a __symbols__ node at the
root node of the resulting blob for any node labels used.

When using the /plugin/ tag all unresolved label references
be tracked in the __fixups__ node, while all local phandle
references will the tracked in the __local_fixups__ node.

This is sufficient to implement a dynamic DT object loader.

Signed-off-by: Pantelis Antoniou <panto@antoniou-consulting.com>
[hvaibhav@ti.com: Patch available at patchwork doesn't apply cleanly
    so using updated commit from "pantalis/not-capebus-v28" branch.]
Signed-off-by: Vaibhav Hiremath <hvaibhav@ti.com>