If the parameter buffer is small enough, try to allocate it with kmalloc() rather than vmalloc().
vmalloc is noticeably slower than kmalloc because it has to manipulate
page tables.

On PA-RISC this patch speeds up activation 13 times.
On Opteron this patch speeds up activation by 5%.

This patch introduces a new function free_params() to free the
parameters and this uses new flags that record whether or not vmalloc()
was used and whether or not the input buffer must be wiped after use.

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Alasdair G Kergon <agk@redhat.com>